fakeTranspositFunctions to transpositFunctions, transpositFunctions to realTranspositFunctions
