+++ src/transpositFunctions.ts
@@ -1,111 +1,28 @@
-import moment from "moment";
-import {Transposit, User} from "transposit";
 import {Car, CarEvents, CarRequest} from "./types";
 import * as React from "react";
 
-export const transposit = new Transposit(
-    "https://lmc-carshare-89gbj.transposit.io"
-);
+interface User {
+    "name": string,
+    "email": string
+}
 
 interface AvailableCars {
     "start": string;
     "end": string;
-    "cars": Car[];
+    "cars": string[];
 }
 
-interface TranspositResponse {
+export interface TranspositResponse {
     error: string;
     response: any;
 }
 
-export async function getAvailableCars(pendingRequest :CarRequest) {
+export async function listRequests() {
     let response = {
         error: "",
-        response: [] as Car[]
+        response: fakeRequests as CarRequest[]
     };
-    await transposit
-        .run("get_cars_available_for_time", {
-            start: pendingRequest.start,
-            end: pendingRequest.end
-        })
-        .then(results => {
-            let selectedFeatures = pendingRequest.features;
-            let raw_cars = results.results[0] as AvailableCars;
-            response.response = raw_cars.cars.filter(x => {
-                let res = true;
-                for (let i in selectedFeatures) {
-                    if (x.Features.indexOf(selectedFeatures[i]) < 0) {
-                        res = false;
-                    }
-                }
-                return res;
-            });
-            return response;
-        })
-        .catch(response => {
-            response.error = response.toString();
-            return response;
-        });
-}
-
-export async function deleteRequests(eventIds :string[]) {
-    let response = {
-        error: "",
-        response: [] as CarRequest[]
-    };
-    transposit
-        .run("delete_requests", {
-            eventIds: eventIds.toString()
-        })
-        .then(x => {
-            response.response = x.results as CarRequest[];
-            return response;
-        })
-        .catch(response => {
-            response.error = response.toString();
-            return response;
-        });
-}
-
-export async function sendReminderToOwner(eventId: string) {
-    let response = {
-        error: "",
-        response: null
-    };
-    await transposit
-        .run("send_reminder", {
-            eventId: eventId
-        })
-        .then(x => {
-            console.log(x);
-            return response;
-        })
-        .catch(response => {
-            response.error = response.toString();
-            return response;
-        });
-}
-
-export async function createBooking(pendingRequest: CarRequest) {
-    let response = {
-        error: "",
-        response: null as CarRequest | null
-    };
-    await transposit
-        .run("create_reservation", {
-            start: pendingRequest.start,
-            end: pendingRequest.end,
-            requester: pendingRequest.requester,
-            vehicle: pendingRequest.vehicle
-        })
-        .then(results => {
-            response.response = results.results[0] as CarRequest;
-            return response;
-        })
-        .catch(results => {
-            response.error = results.toString();
-            return response;
-        });
+    return Promise.resolve(response);
 }
 
 export async function listFeatures() {
@@ -113,73 +30,78 @@ export async function listFeatures() {
         error: "",
         response: [] as string[]
     };
-    await transposit
-        .run("list_features", {})
-        .then(x => {
-            console.log("features listed");
-            response.response = x.results as string[];
-            return response;
-        })
-        .catch(x => {
-            console.log("features error");
-            response.error = x.toString();
-            return response;
-        });
+    response.response = [
+        "pet friendly",
+        "child friendly",
+        "eco friendly",
+        "cargo",
+        "camping"
+    ];
+    return Promise.resolve(response);
 }
 
 export async function listAllCars() {
     let response = {
         error: "",
-        response: [] as Car[]
+        response: fakeCars as Car[]
     };
-    await transposit
-        .run("list_cars", {})
-        .then(x => {
-            response.response = Object.keys(x.results[0]).map(key => x.results[0][key]) as Car[];
-            return response;
-        })
-        .catch(x => {
-            response.error = x.toString();
-            return response;
-        });
-    return response;
+    return Promise.resolve(response);
+}
+
+export async function getAvailableCars(pendingRequest :CarRequest) {
+    let fakeResponse = {
+        start: pendingRequest.start,
+        end: pendingRequest.end,
+        cars: [fakeCars[1].Licence]
+    } as AvailableCars;
+    let response = {
+        error: "",
+        response: fakeResponse.cars
+    };
+    return Promise.resolve(response);
+}
+
+export async function deleteRequests(eventIds :string[]) {
+    let reqs = fakeRequests;
+    let deleted = [] as CarRequest[];
+    console.log(`deleteRequests ${eventIds}`);
+    let eventMap = reqs.map(y => {return y.eventId;});
+    eventIds.forEach(x => {
+        let index = eventMap.indexOf(x);
+        if (index >= 0) {
+            deleted.push(fakeRequests[index] as CarRequest);
+        }
+    });
+    let response = {
+        error: "",
+        response: deleted
+    };
+    return Promise.resolve(response);
+}
+
+export async function sendReminderToOwner(eventId: string) {
+    let response = {
+        error: "",
+        response: null
+    };
+    return Promise.resolve(response);
+}
+
+export async function createBooking(pendingRequest: CarRequest) {
+    console.log(`createBooking with ${pendingRequest.vehicle}`);
+    let response = {
+        error: "",
+        response: pendingRequest as CarRequest | null
+    };
+    return Promise.resolve(response);
 }
 
 export async function getThreeDaysEvents() {
     let response = {
         error: "",
-        response: null as CarEvents | null
+        response: fakeCarEvent
     };
-    await transposit
-        .run("three_day_array", {start: moment().format(), interval: '900'})
-        .then(x => {
-            response.response = x.results[0] as CarEvents
-            return response;
-        })
-        .catch(x => {
-            response.error = x.toString();
-            return response;
-        });
-}
-
-export async function listRequests() {
-    let response = {
-        error: "",
-        response: [] as CarRequest[]
-    };
-    transposit
-        .run("list_requests", {})
-        .then(results => {
-            // shift off the labels
-            results.results.shift();
-            let requests = results.results as CarRequest[];
-            response.response = requests;
-            return response;
-        })
-        .catch(x => {
-            response.error = x.toString();
-            return response;
-        });
+    return Promise.resolve(response);
 }
 
 
@@ -189,10 +111,6 @@ export async function listRequests() {
 export function useSignedIn(): boolean {
     const [isSignedIn, setIsSignedIn] = React.useState<boolean>(false);
     React.useEffect(() => {
-        if (!transposit.isSignedIn()) {
-            window.location.href = "/signin";
-            return;
-        }
         setIsSignedIn(true);
     }, []);
     return isSignedIn;
@@ -207,10 +125,7 @@ export function useUser(isSignedIn: boolean): User | null {
         if (!isSignedIn) {
             return;
         }
-        transposit
-            .loadUser()
-            .then(u => setUser(u))
-            .catch(response => console.log(response));
+    setUser({name: "Testy McTest", email: "test@test.com"})
     }, [isSignedIn]);
     return user;
 }
@@ -219,19 +134,7 @@ export function useIsValidMember(user: User | null): number {
     const [isValid, setValid] = React.useState<number>(0);
     React.useEffect(() => {
         if (user) {
-            transposit
-                .run("is_valid_member", {email: user.email})
-                .then(x => {
-                    if (x.results[0]) {
-                        setValid(1);
-                    } else {
-                        setValid(-1);
-                    }
-                })
-                .catch(response => {
-                    console.log(response.toString());
-                    setValid(-1);
-                });
+            setValid(1);
         }
     }, [user, isValid]);
 
@@ -242,20 +145,10 @@ export function useIsValidMember(user: User | null): number {
  * Handle sign-in page
  */
 export async function signIn() {
-    await transposit.signIn(
-        `${window.location.origin}/signin/handle-redirect`
-    );
 }
 
 export function SignInHandleRedirect() {
-    transposit.handleSignIn().then(
-        () => {
-            window.location.href = "/bookings";
-        },
-        () => {
-            window.location.href = "/signin";
-        }
-    );
+    window.location.href = "/bookings";
     return null;
 }
 
@@ -263,6 +156,93 @@ export function SignInHandleRedirect() {
  * Handle sign-out page
  */
 export function SignOut() {
-    transposit.signOut(`${window.location.origin}/signin`);
     return null;
 }
+
+const fakeRequests = [
+    {
+        "threadId": "175916c0d4bbe0ec",
+        "vehicle": "AL675T",
+        "requester": "pwcottle@gmail.com",
+        "start": "2021-01-01T04:00:00-0800",
+        "end": "2021-01-01T13:00:00-0800",
+        "eventId": "qs7eb1proc33bgal7jrv431rak",
+        "confirmed": "TRUE"
+    },
+    {
+        "threadId": "1759168e6f62e313",
+        "vehicle": "NLEAF",
+        "requester": "pwcottle@gmail.com",
+        "start": "2021-03-10T05:00:00-0800",
+        "end": "2021-03-10T14:00:00-0800",
+        "eventId": "o106pbpmlcap5t6a4oc16b335g",
+        "confirmed": "TRUE"
+    }
+];
+
+const fakeCars = [
+    {
+        "Timestamp": "20/01/2020 16:21:11",
+        "Make": "Toyota",
+        "Model": "Prius",
+        "Color": "Blue",
+        "Features": [
+            "pet friendly",
+            "child friendly",
+            "eco friendly"
+        ],
+        "Licence": "AL675T",
+        "Email": "lmc.blue.prius.2009@gmail.com",
+        "AlwaysAvailable": true,
+        "Confirm": false,
+        "Description": "Blue Toyota Prius AL675T"
+    },
+    {
+        "Timestamp": "20/01/2020 16:20:52",
+        "Make": "Nissan",
+        "Model": "Leaf",
+        "Color": "Orange",
+        "Features": [
+            "child friendly",
+            "eco friendly"
+        ],
+        "Licence": "NLEAF",
+        "Email": "lmc.orange.leaf.2017@gmail.com",
+        "AlwaysAvailable": true,
+        "Confirm": true,
+        "Description": "Orange Nissan Leaf NLEAF"
+    },
+    {
+        "Timestamp": "20/01/2020 16:24:21",
+        "Make": "Honda",
+        "Model": "Element",
+        "Color": "Orange",
+        "Features": [
+            "pet friendly",
+            "cargo",
+            "camping"
+        ],
+        "Licence": "ELEMENT",
+        "Email": "mutantdaisies@gmail.com",
+        "AlwaysAvailable": false,
+        "Confirm": true,
+        "Description": "Orange Honda Element ELEMENT"
+    }
+];
+
+const fakeCarEvent =
+    {
+        "start": "2020-11-20T00:00:00-08:00",
+        "end": "2020-11-23T00:00:00-08:00",
+        "cars": [
+            "AL675T",
+            "NLEAF",
+            "ELEMENT"
+        ],
+        "interval": 900,
+        "busy_segments": [
+            ",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,<-->,,,,,,,,,,,,,,,,",
+            ",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,",
+            ",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,"
+        ]
+    } as CarEvents;
